TS-35130 Update jacoco agent
