+++ agent/src/main/java/com/teamscale/jacoco/agent/AgentBase.java
@@ -1,17 +1,7 @@
 package com.teamscale.jacoco.agent;
 
-import com.teamscale.client.HttpUtils;
-import com.teamscale.jacoco.agent.options.AgentOptionParseException;
 import com.teamscale.jacoco.agent.options.AgentOptions;
-import com.teamscale.jacoco.agent.options.AgentOptionsParser;
-import com.teamscale.jacoco.agent.options.FilePatternResolver;
-import com.teamscale.jacoco.agent.options.JacocoAgentBuilder;
-import com.teamscale.jacoco.agent.util.DebugLogDirectoryPropertyDefiner;
-import com.teamscale.jacoco.agent.util.LogDirectoryPropertyDefiner;
 import com.teamscale.jacoco.agent.util.LoggingUtils;
-import com.teamscale.jacoco.agent.util.LoggingUtils.LoggingResources;
-import org.conqat.lib.commons.collections.CollectionUtils;
-import org.conqat.lib.commons.filesystem.FileSystemUtils;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.ServerConnector;
 import org.eclipse.jetty.servlet.ServletContextHandler;
@@ -22,15 +12,7 @@ import org.glassfish.jersey.servlet.ServletContainer;
 import org.jacoco.agent.rt.RT;
 import org.slf4j.Logger;
 
-import java.io.File;
-import java.io.IOException;
-import java.lang.instrument.Instrumentation;
 import java.lang.management.ManagementFactory;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.List;
-import java.util.Optional;
 
 /**
  * Base class for agent implementations. Handles logger shutdown, store creation and instantiation of the
@@ -50,8 +32,6 @@ public abstract class AgentBase {
 	protected AgentOptions options;
 
 	private Server server;
-
-	private static LoggingUtils.LoggingResources loggingResources;
 
 	/** Constructor. */
 	public AgentBase(AgentOptions options) throws IllegalStateException {
@@ -128,133 +108,14 @@ public abstract class AgentBase {
 	protected abstract ResourceConfig initResourceConfig();
 
 	/**
-	 * Called by the actual premain method once the agent is isolated from the rest of the application.
-	 */
-	public static void premain(String options, Instrumentation instrumentation) throws Exception {
-		AgentOptions agentOptions;
-		DelayedLogger delayedLogger = new DelayedLogger();
-
-		List<String> javaAgents = CollectionUtils.filter(ManagementFactory.getRuntimeMXBean().getInputArguments(),
-				s -> s.contains("-javaagent"));
-		if (javaAgents.size() > 1) {
-			delayedLogger.warn("Using multiple java agents could interfere with coverage recording.");
-		}
-		if (!javaAgents.get(0).contains("teamscale-jacoco-agent.jar")) {
-			delayedLogger.warn("For best results consider registering the Teamscale JaCoCo Agent first.");
-		}
-
-		try {
-			agentOptions = AgentOptionsParser.parse(options, delayedLogger);
-		} catch (AgentOptionParseException e) {
-			try (LoggingUtils.LoggingResources ignored = initializeFallbackLogging(options, delayedLogger)) {
-				delayedLogger.error("Failed to parse agent options: " + e.getMessage(), e);
-				System.err.println("Failed to parse agent options: " + e.getMessage());
-
-				// we perform actual logging output after writing to console to
-				// ensure the console is reached even in case of logging issues
-				// (see TS-23151). We use the Agent class here (same as below)
-				Logger logger = LoggingUtils.getLogger(Agent.class);
-				delayedLogger.logTo(logger);
-				throw e;
-			}
-		}
-
-		initializeLogging(agentOptions, delayedLogger);
-		Logger logger = LoggingUtils.getLogger(Agent.class);
-		delayedLogger.logTo(logger);
-
-		HttpUtils.setShouldValidateSsl(agentOptions.shouldValidateSsl());
-
-		logger.info("Starting JaCoCo's agent");
-		JacocoAgentBuilder agentBuilder = new JacocoAgentBuilder(agentOptions);
-		JaCoCoPreMain.premain(agentBuilder.createJacocoAgentOptions(), instrumentation, logger);
-
-		AgentBase agent = agentBuilder.createAgent(instrumentation);
-		agent.registerShutdownHook();
-	}
-
-	/** Initializes logging during {@link #premain(String, Instrumentation)} and also logs the log directory. */
-	private static void initializeLogging(AgentOptions agentOptions, DelayedLogger logger) throws IOException {
-		if (agentOptions.isDebugLogging()) {
-			initializeDebugLogging(agentOptions, logger);
-		} else {
-			loggingResources = LoggingUtils.initializeLogging(agentOptions.getLoggingConfig());
-			logger.info("Logging to " + new LogDirectoryPropertyDefiner().getPropertyValue());
-		}
-	}
-
-	/**
-	 * Initializes debug logging during {@link #premain(String, Instrumentation)} and also logs the log directory if
-	 * given.
-	 */
-	private static void initializeDebugLogging(AgentOptions agentOptions, DelayedLogger logger) {
-		loggingResources = LoggingUtils.initializeDebugLogging(agentOptions.getDebugLogDirectory());
-		Path logDirectory = Paths.get(new DebugLogDirectoryPropertyDefiner().getPropertyValue());
-		if (FileSystemUtils.isValidPath(logDirectory.toString()) && Files.isWritable(logDirectory)) {
-			logger.info("Logging to " + logDirectory);
-		} else {
-			logger.warn("Could not create " + logDirectory + ". Logging to console only.");
-		}
-	}
-
-	/**
-	 * Initializes fallback logging in case of an error during the parsing of the options to
-	 * {@link #premain(String, Instrumentation)} (see TS-23151). This tries to extract the logging configuration and use
-	 * this and falls back to the default logger.
-	 */
-	private static LoggingResources initializeFallbackLogging(String premainOptions, DelayedLogger delayedLogger) {
-		for (String optionPart : premainOptions.split(",")) {
-			if (optionPart.startsWith(AgentOptionsParser.LOGGING_CONFIG_OPTION + "=")) {
-				return createFallbackLoggerFromConfig(optionPart.split("=", 2)[1], delayedLogger);
-			}
-
-			if (optionPart.startsWith(AgentOptionsParser.CONFIG_FILE_OPTION + "=")) {
-				String configFileValue = optionPart.split("=", 2)[1];
-				Optional<String> loggingConfigLine = Optional.empty();
-				try {
-					File configFile = new FilePatternResolver(delayedLogger).parsePath(
-							AgentOptionsParser.CONFIG_FILE_OPTION, configFileValue).toFile();
-					loggingConfigLine = FileSystemUtils.readLinesUTF8(configFile).stream()
-							.filter(line -> line.startsWith(AgentOptionsParser.LOGGING_CONFIG_OPTION + "="))
-							.findFirst();
-				} catch (IOException | AgentOptionParseException e) {
-					delayedLogger.error("Failed to load configuration from " + configFileValue + ": " + e.getMessage(),
-							e);
-				}
-				if (loggingConfigLine.isPresent()) {
-					return createFallbackLoggerFromConfig(loggingConfigLine.get().split("=", 2)[1], delayedLogger);
-				}
-			}
-		}
-
-		return LoggingUtils.initializeDefaultLogging();
-	}
-
-	/** Creates a fallback logger using the given config file. */
-	private static LoggingResources createFallbackLoggerFromConfig(String configLocation, DelayedLogger delayedLogger) {
-		try {
-			return LoggingUtils.initializeLogging(
-					new FilePatternResolver(delayedLogger).parsePath(AgentOptionsParser.LOGGING_CONFIG_OPTION,
-							configLocation));
-		} catch (IOException | AgentOptionParseException e) {
-			String message = "Failed to load log configuration from location " + configLocation + ": " + e.getMessage();
-			delayedLogger.error(message, e);
-			// output the message to console as well, as this might
-			// otherwise not make it to the user
-			System.err.println(message);
-			return LoggingUtils.initializeDefaultLogging();
-		}
-	}
-
-	/**
 	 * Registers a shutdown hook that stops the timer and dumps coverage a final time.
 	 */
-	private void registerShutdownHook() {
+	void registerShutdownHook() {
 		Runtime.getRuntime().addShutdownHook(new Thread(() -> {
 			stopServer();
 			prepareShutdown();
 			logger.info("CQSE JaCoCo agent successfully shut down.");
-			loggingResources.close();
+			PreMain.closeLoggingResources();
 		}));
 	}
