+++ agent/src/main/java/com/teamscale/jacoco/agent/testimpact/TestwiseCoverageAgent.java
@@ -7,9 +7,13 @@ package com.teamscale.jacoco.agent.testimpact;
 
 import com.teamscale.jacoco.agent.AgentBase;
 import com.teamscale.jacoco.agent.options.AgentOptions;
+import com.teamscale.jacoco.agent.util.LoggingUtils;
 import com.teamscale.report.testwise.jacoco.JaCoCoTestwiseReportGenerator;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.server.ServerProperties;
+import org.slf4j.Logger;
+
+import java.io.IOException;
 
 /**
  * A wrapper around the JaCoCo Java agent that starts a HTTP server and listens for test events.
@@ -20,6 +24,17 @@ public class TestwiseCoverageAgent extends AgentBase {
 	 * The test event strategy handler.
 	 */
 	protected final TestEventHandlerStrategyBase testEventHandler;
+
+	/** Creates a {@link TestwiseCoverageAgent} based on the given options. */
+	public static TestwiseCoverageAgent create(AgentOptions agentOptions) throws IOException {
+		Logger logger = LoggingUtils.getLogger(JaCoCoTestwiseReportGenerator.class);
+		JaCoCoTestwiseReportGenerator reportGenerator = new JaCoCoTestwiseReportGenerator(
+				agentOptions.getClassDirectoriesOrZips(), agentOptions.getLocationIncludeFilter(),
+				agentOptions.getDuplicateClassFileBehavior(), LoggingUtils.wrap(logger));
+		return new TestwiseCoverageAgent(agentOptions,
+				new TestExecutionWriter(agentOptions.createNewFileInOutputDirectory("test-execution", "json")),
+				reportGenerator);
+	}
 
 
 	public TestwiseCoverageAgent(AgentOptions options, TestExecutionWriter testExecutionWriter,
