+++ agent/src/main/resources/com/teamscale/jacoco/agent/logback-default-debugging.xml
@@ -2,24 +2,24 @@
 <configuration>
 	<define name="defaultLogDir" class="com.teamscale.jacoco.agent.util.DebugLogDirectoryPropertyDefiner" />
 
-	<appender name="CONSOLE" class="shadow.ch.qos.logback.core.ConsoleAppender">
+	<appender name="CONSOLE" class="tsshadow.ch.qos.logback.core.ConsoleAppender">
 		<encoder>
 			<pattern>%date{yyyy-MM-dd HH:mm:ss} %-5level %logger{35} - %msg%n</pattern>
 		</encoder>
 	</appender>
 
-	<appender name="RollingFile" class="shadow.ch.qos.logback.core.rolling.RollingFileAppender">
+	<appender name="RollingFile" class="tsshadow.ch.qos.logback.core.rolling.RollingFileAppender">
 		<!-- TODO: adjust the path where the log file should be written -->
 		<file>${defaultLogDir}/teamscale-jacoco-agent.log</file>
 
-		<rollingPolicy class="shadow.ch.qos.logback.core.rolling.FixedWindowRollingPolicy">
+		<rollingPolicy class="tsshadow.ch.qos.logback.core.rolling.FixedWindowRollingPolicy">
 			<!-- TODO: adjust the path where the log file should be written -->
 			<fileNamePattern>${defaultLogDir}/teamscale-jacoco-agent-%i.log.zip</fileNamePattern>
 			<minIndex>1</minIndex>
 			<maxIndex>10</maxIndex>
 		</rollingPolicy>
 
-		<triggeringPolicy class="shadow.ch.qos.logback.core.rolling.SizeBasedTriggeringPolicy">
+		<triggeringPolicy class="tsshadow.ch.qos.logback.core.rolling.SizeBasedTriggeringPolicy">
 			<maxFileSize>1MB</maxFileSize>
 		</triggeringPolicy>
 
@@ -33,6 +33,6 @@
 		<appender-ref ref="CONSOLE"/>
 	</root>
 
-	<logger name="shadow.org.apache.spark" level="warn" />
-	<logger name="shadow.org.eclipse.jetty" level="warn" />
+	<logger name="tsshadow.org.apache.spark" level="warn" />
+	<logger name="tsshadow.org.eclipse.jetty" level="warn" />
 </configuration>
