+++ agent/src/main/resources/com/teamscale/jacoco/agent/logback-default.xml
@@ -2,17 +2,17 @@
 <configuration>
     <define name="defaultLogDir" class="com.teamscale.jacoco.agent.util.LogDirectoryPropertyDefiner" />
 
-    <appender name="RollingFile" class="shadow.ch.qos.logback.core.rolling.RollingFileAppender">
+    <appender name="RollingFile" class="tsshadow.ch.qos.logback.core.rolling.RollingFileAppender">
         <!-- We write to a file that is located in a folder provided via the context -->
         <file>${defaultLogDir}/teamscale-jacoco-agent.log</file>
 
-        <rollingPolicy class="shadow.ch.qos.logback.core.rolling.FixedWindowRollingPolicy">
+        <rollingPolicy class="tsshadow.ch.qos.logback.core.rolling.FixedWindowRollingPolicy">
             <fileNamePattern>${defaultLogDir}/teamscale-jacoco-agent-%i.log.zip</fileNamePattern>
             <minIndex>1</minIndex>
             <maxIndex>10</maxIndex>
         </rollingPolicy>
 
-        <triggeringPolicy class="shadow.ch.qos.logback.core.rolling.SizeBasedTriggeringPolicy">
+        <triggeringPolicy class="tsshadow.ch.qos.logback.core.rolling.SizeBasedTriggeringPolicy">
             <maxFileSize>1MB</maxFileSize>
         </triggeringPolicy>
 
@@ -25,6 +25,6 @@
         <appender-ref ref="RollingFile"/>
     </root>
 
-    <logger name="shadow.org.apache.spark" level="warn" />
-    <logger name="shadow.org.eclipse.jetty" level="warn" />
+    <logger name="tsshadow.org.apache.spark" level="warn" />
+    <logger name="tsshadow.org.eclipse.jetty" level="warn" />
 </configuration>
