+++ gradle/libs.versions.toml
@@ -2,8 +2,9 @@
 jetty = "9.4.52.v20230823"
 jersey = "2.40"
 # When upgrading JaCoCo to a newer version make sure to
-# check the comment in the OpenAnalyzer.java, JaCoCoPreMain.java and CachingInstructionsBuilder.java.
-jacoco = "0.8.8"
+# check the comment in the OpenAnalyzer.java, JaCoCoPreMain.java and CachingInstructionsBuilder.java
+# and update the internal_xxxxxx hash included in the imports in LenientCoverageTransformer.java.
+jacoco = "0.8.10"
 # We need to stay on the 1.3.x release line as 1.4.x requires Java 11
 logback = "1.3.11"
 retrofit = "2.9.0"
