+++ system-tests/default-excludes-test/src/main/java/tsshadow/ShadowedClass.java
@@ -1,4 +1,4 @@
-package shadow;
+package tsshadow;
 
 public class ShadowedClass {
 	public int x() {
